Fix remaining 39 errorprone warnings in main sources
  - Suppress EmptyCatch on intentionally empty catch blocks (8 files)
  - Fix javadoc tags: {@param x} -> {@code x}, {@see X} -> {@link X},
    {@link this#method} -> {@link #method}, remove invalid {@inheritDoc}
  - Remove unused methods/constructors (ClientContextImpl, HttpException,
    FrontendResourceImpl) and unused variable (DefaultExceptionMapper)
  - Suppress ThreadPriorityCheck, JavaLangClash, JavaUtilDate,
    ArrayRecordComponent, TypeParameterUnusedInFormals, HidingField
  - Fix OverrideThrowableToString: toString() -> getMessage() in BatchedException
  - Fix NonOverridingEquals: rename equals(Asset) -> contentEquals(Asset)
  - Fix UnnecessaryLambda: convert lambda constant to method reference
  - Fix AnnotateFormatMethod: add @FormatMethod to ValidationError constructor
  - Fix FutureReturnValueIgnored with var ignored pattern
  - Fix InconsistentCapitalization, NonCanonicalType, VariableNameSameAsType,
    StringSplitter, AssignmentExpression, BooleanLiteral, NarrowCalculation,
    NarrowingCompoundAssignment, InlineFormatString, EmptyTopLevelDeclaration,
    ClassCanBeStatic

+++ hivemq-edge/src/main/java/com/hivemq/limitation/TopicAliasLimiterImpl.java
@@ -55,7 +55,7 @@ public class TopicAliasLimiterImpl implements TopicAliasLimiter {
     @Override
     public void initUsage(final int size) {
         // 4 bytes per topic as index
-        this.memoryUsage.addAndGet(size * 4);
+        this.memoryUsage.addAndGet((long) size * 4);
     }
 
     @Override
@@ -68,7 +68,7 @@ public class TopicAliasLimiterImpl implements TopicAliasLimiter {
     public void removeUsage(final String... topics) {
         for (final String topic : topics) {
             if (topic != null) {
-                this.memoryUsage.addAndGet(-1 * getEstimatedSize(topic));
+                this.memoryUsage.addAndGet(-1L * getEstimatedSize(topic));
                 this.topicAliasesTotal.decrementAndGet();
             }
         }
@@ -77,7 +77,7 @@ public class TopicAliasLimiterImpl implements TopicAliasLimiter {
     @Override
     public void finishUsage(final @NotNull String... topics) {
         // 4 bytes per topic as index
-        this.memoryUsage.addAndGet(topics.length * -4);
+        this.memoryUsage.addAndGet((long) topics.length * -4);
         this.removeUsage(topics);
     }
